Education Page & Styling
Added two sections to the top above the table, removed an idea that did not validate from the tables, added a link at the bottom that reloads the page to the top. I was too confused on the java script to make it a scroll up button like I originally thought about. Worked on the styling as well to make it look how I wanted it to.